Enhanced movement a bit. Added ranged enemies. Started on health
Enemies will now enter a wander mode when the target is too far. Ranged enemies will wander away if the target approaches them. Ranged enemies shoot a projectile in a straight line towards the target. Enemies have health but all that is there right now is to be destroyed when it is at 0. Need player to be coded to continue.