Added basic navigation
The enemy will not go into obstacles. Have not tried making them go around yet but they will not step on obstacles. Would be best if we made the walls and traps work like the example obstacle I have as the navmesh will carve around them even if they are moved. Rotated everything to be on the  x and z axis as navmesh is not supported in 2D.